Reorganize AWS tests,
