+++ .github/workflows/check and test.yml
@@ -57,5 +57,5 @@ jobs:
       env:
         PYTHONDEVMODE: 1 # see https://docs.python.org/3/library/devmode.html#devmode
         GRPC_VERBOSITY: debug
-      run: | # please keep in alphabetical order
-        poetry run pytest tests/test_conda.py tests/test_func_worker.py tests/test_local_cache.py tests/test_local_code.py tests/test_object_storage.py tests/test_task_result.py
+      run: |
+        poetry run pytest tests/test_conda.py tests/test_func_worker.py tests/test_instance_allocation.py tests/test_local_cache.py tests/test_local_code.py tests/test_object_storage.py tests/test_task_result.py
